requirements.txt
Flask==1.0.3
# https://download.pytorch.org/whl/cpu/torch-1.0.0-cp37-cp37m-linux_x86_64.whl
torchvision==0.2.1
numpy==1.16.4
Pillow==7.1.0
gunicorn==19.9.0
pytorch-lightning==1.0.8
torch==1.7.0
torchtext==0.3.1
torchvision==0.8.1
